Setup language mode context
