+++ src/contexts/LanguageContext.tsx
@@ -1,0 +1,3 @@
+import { createContext } from 'react';
+
+export const LanguageModeContext = createContext<'polish' | 'english'>('polish');
