+++ src/index.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+import ReactDOM from 'react-dom/client';
+import MainProvider from './MainProvider';
+
+ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(<MainProvider />);
