--- src/main.tsx
@@ -1,19 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import HomePage from './components/HomePage';
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import { HOME_PATH } from './constants/RouterConstants';
-import './styles/Constants.scss';
-
-const router = createBrowserRouter([
-  {
-    path: HOME_PATH,
-    element: <HomePage />,
-  },
-]);
-
-ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-  <React.StrictMode>
-    <RouterProvider router={router} />
-  </React.StrictMode>
-);
